+++ examples/demo_platformer/debug.rs
@@ -1,4 +1,7 @@
-use bevy::{input::common_conditions::input_toggle_active, prelude::*};
+use bevy::{
+    dev_tools::diagnostics_overlay::*, diagnostic::FrameTimeDiagnosticsPlugin,
+    input::common_conditions::input_toggle_active, prelude::*,
+};
 
 use bevy_inspector_egui::{
     bevy_egui::{EguiGlobalSettings, EguiPlugin},
@@ -17,21 +20,30 @@ pub(super) fn plugin(app: &mut App) {
         EguiPlugin::default(),
         WorldInspectorPlugin::default().run_if(input_toggle_active(false, TOGGLE_INSPECTOR_KEY)),
         avian2d::prelude::PhysicsDebugPlugin,
+        DiagnosticsOverlayPlugin,
+        FrameTimeDiagnosticsPlugin::default(),
     ));
 
-    app.add_systems(Startup, setup_help_text);
+    app.add_systems(Startup, (setup_help_text, setup_diagnostics_overlay));
 }
 
 fn setup_help_text(mut commands: Commands) {
     commands
-        .spawn(Node {
-            display: Display::Flex,
-            align_self: AlignSelf::FlexEnd,
-            flex_direction: FlexDirection::Column,
-            ..default()
-        })
+        .spawn((
+            Name::new("Helper text"),
+            Node {
+                display: Display::Flex,
+                align_self: AlignSelf::FlexEnd,
+                flex_direction: FlexDirection::Column,
+                ..default()
+            },
+        ))
         .with_children(|builder| {
             builder.spawn(Text(String::from("Debug mode enabled")));
             builder.spawn(Text(String::from("Toggle inspector: [F1]")));
         });
+}
+
+fn setup_diagnostics_overlay(mut commands: Commands) {
+    commands.spawn((Name::new("Diagnostic overlay"), DiagnosticsOverlay::fps()));
 }