+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [unreleased]
 
+### Features
+
+- add support for list properties (#212)
+
+### Features
+
+- Update tiled-rs v0.16
+- Add support for capsule objects
+- Allow self-referential for physics colliders
+- Use gizmos for rendering TiledDebugTilesPlugin text
+
 ## v0.13.1
 
 ### Features